% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{common_data}
\alias{common_data}
\title{Pivot common branches to wide format (internal)}
\usage{
common_data(complete_data)
}
\arguments{
\item{complete_data}{`data.frame` from [get_complete_data()].}
}
\value{
Wide `data.frame` with one row per package and columns for
  `PROD`, `DEV_v2`, `QA` branch versions.
}
\description{
Pivot common branches to wide format (internal)
}
\keyword{internal}
