% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{get_package_version}
\alias{get_package_version}
\title{Get DESCRIPTION versions for all branches of each package (internal)}
\usage{
get_package_version(package)
}
\arguments{
\item{package}{Character vector of core PIP package names.}
}
\value{
Named list where each element corresponds to a package and
  contains a named character vector of branch -> version mappings.
}
\description{
Get DESCRIPTION versions for all branches of each package (internal)
}
\keyword{internal}
