% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{join_and_get_status}
\alias{join_and_get_status}
\title{Join local install info with remote and compute status (internal)}
\usage{
join_and_get_status(local, dev, branch_to_compare)
}
\arguments{
\item{local}{`data.frame` with columns `package`, `local_branch`,
`local_version`.}

\item{dev}{`data.frame` with columns `package`, `branch`, `version`
for the comparison branch.}

\item{branch_to_compare}{Character scalar. Name of the comparison
branch (for labelling).}
}
\value{
`data.frame` with columns: `package`, `local_branch`,
  `local_version`, `local_status`.
}
\description{
Join local install info with remote and compute status (internal)
}
\keyword{internal}
