% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapip-package.R, R/metapip.R
\docType{package}
\name{metapip}
\alias{metapip}
\title{metapip: PIP packages manager}
\description{
The goal of metapip is to provide the user with a set of functions that
allow him/her to work efficiently with all the PIP R packages. The technical
world of PIP consist of several packages that interact to each to load,
format, modify, and estimate data for the PIP system. Given that the number
of packages has increased over time, it was necessary to create a meta
package whose only objective is the proper management of all the other PIP
packages.
}
\section{metapip functions}{
 The metapip functions ...
}

