% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowname_to_column}
\alias{rowname_to_column}
\title{Move row names to a column}
\usage{
rowname_to_column(data, var)
}
\arguments{
\item{data}{A data.frame.}

\item{var}{Character scalar. Name of the new column to hold row names.}
}
\value{
A data.frame with the same columns as `data` plus the new
  column `var` as the first column. Row names are removed.
}
\description{
A lightweight alternative to `tibble::rownames_to_column()` that does
not require the tibble package. Prepends the row names of a data.frame
as a new column.
}
\examples{
df <- data.frame(x = 1:3, row.names = c("a", "b", "c"))
rowname_to_column(df, "id")

}
