% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rs_theme}
\alias{rs_theme}
\title{Detect RStudio theme for colorDF display settings (internal)}
\usage{
rs_theme()
}
\value{
A list with elements `editor`, `global`, `dark`,
  `foreground`, `background`. Returns an empty template when
  RStudio is not available.
}
\description{
Detect RStudio theme for colorDF display settings (internal)
}
\keyword{internal}
