% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_metapip.R
\name{set_custom_default_branch}
\alias{set_custom_default_branch}
\title{Set default custom branching options}
\usage{
set_custom_default_branch(...)
}
\arguments{
\item{...}{Named elements to be added or updated in the custom default list.}
}
\value{
invisible NULL
}
\description{
Set default custom branching options
}
\examples{
{
set_custom_default_branch(pipr = 'main', 'pipapi' = 'DEV_v3')
}

}
