% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{split_packages_into_list}
\alias{split_packages_into_list}
\title{Split non-common branches into per-package list (internal)}
\usage{
split_packages_into_list(complete_data)
}
\arguments{
\item{complete_data}{`data.frame` from [get_complete_data()].}
}
\value{
Named list of `data.frame`s, one per package, each with
  columns `branch` and `version`.
}
\description{
Split non-common branches into per-package list (internal)
}
\keyword{internal}
