% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_table.R
\name{status_table}
\alias{status_table}
\title{Status tables of package versions in different branches
Apart from PROD and DEV, we are only showing those branches whose version is higher than DEV.}
\usage{
status_table(package = NULL)
}
\arguments{
\item{package}{One (or more) of the PIP core packages. Default NULL will include all the packages}
}
\value{
tibble of pipr packages and the corresponding package versions of branch
}
\description{
Status tables of package versions in different branches
Apart from PROD and DEV, we are only showing those branches whose version is higher than DEV.
}
