% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_github_token}
\alias{check_github_token}
\title{make sure your GITHUB credentials are properly setup}
\usage{
check_github_token()
}
\value{
invisible TRUE if credentials are perfectly set
}
\description{
make sure your GITHUB credentials are properly setup
}
\examples{
\dontrun{
check_github_token()
}
}
