% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_branch_info}
\alias{get_branch_info}
\title{Get last update time of branches in a specific package}
\usage{
get_branch_info(package = "pipapi", branch = NULL)
}
\arguments{
\item{package}{one of the core package (default "pipapi")}

\item{branch}{valid branch name for the specified package}
}
\value{
a knitr::kable output with details about each branch of the package
}
\description{
Get last update time of branches in a specific package
}
\examples{
\dontrun{
get_branch_info()
get_branch_info(branch = c("PROD", "QA"))
get_branch_info(package = "wbpip", branch = c("PROD", "QA"))
}
}
