% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_branches}
\alias{get_branches}
\title{Get available branches for a particular package}
\usage{
get_branches(package = "pipapi", display = TRUE)
}
\arguments{
\item{package}{one of the core package name (default "pipapi")}

\item{display}{(default TRUE) do you want to display the branches in the console?}
}
\value{
an invisible character vector of branches
}
\description{
Get available branches for a particular package
}
\examples{
\dontrun{
  get_branches()
  get_branches("pipr")
}

}
