% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_metapip.R
\name{get_core_packages}
\alias{get_core_packages}
\title{Get core PIP ecosystem package}
\usage{
get_core_packages(exclude = NULL)
}
\arguments{
\item{exclude}{character: packages to exclude from attaching. if `getwd()` is
one of the core PIP packages, that package will be excluded by default. To
avoid that, set exclude to `NULL`.}
}
\value{
character vector with names of PIP packages
}
\description{
Get core PIP ecosystem package
}
\note{
`get_core_packages` is an alias for [get_core_pagkages()]. The
  `get_core_pagkages` spelling (typo) is retained and deprecated for
  backward compatibility.
}
\examples{
get_core_packages()
get_core_packages(exclude = "pipdata")
}
