% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_current_branches}
\alias{get_current_branches}
\alias{get_package_current_branch}
\title{get the current branches that are meant to be used}
\usage{
get_current_branches(package = NULL, verbose = TRUE)

get_package_current_branch(package)
}
\arguments{
\item{package}{character: vector with name of branches. E.g., c("pipdata",
"pipfaker").
TRUE}

\item{verbose}{logical: whether to display all current branches. Default is}
}
\value{
list with names of packages and branches

named character vector with branches of package
}
\description{
It does not necessarily mean that these are the branches currently installed.
[init_metapip] will notify you if that is the case.
}
\examples{
get_current_branches()
get_package_current_branch(c("pipdata", "pipfaker"))
}
