% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_branch}
\alias{get_default_branch}
\alias{get_current_branches}
\title{Get default branch for a particular package}
\usage{
get_default_branch(package)

get_current_branches()
}
\arguments{
\item{package}{One of pip core packages}
}
\value{
name of branch

list with names of packages and branches
}
\description{
It does not necessarily mean that these are the branches currently installed.
[init_metapip] will notify you if that is the case.
}
\examples{
get_default_branch("pipapi")
get_current_branches()
}
