% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_default_branch}
\alias{get_default_branch}
\title{Get default branch in PIP ecosystem}
\usage{
get_default_branch()
}
\value{
name of branch
}
\description{
Get default branch in PIP ecosystem
}
\examples{
get_default_branch()
}
