% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_latest_branch_update}
\alias{get_latest_branch_update}
\title{Get details of the branch which was last updated}
\usage{
get_latest_branch_update(package = "pipapi", display = TRUE)
}
\arguments{
\item{package}{one of the core package (default "pipapi")}

\item{display}{(default TRUE) do you want to display the branches in the console?}
}
\value{
colorDF::colorDF output along with an invisible single row dataframe.
  Timestamps (`last_update_time`) are returned as POSIXct in UTC. If `package`
  has only `gh-pages` branches, returns a single-row data.frame with `NA`
  values (and warns).
}
\description{
This function is useful to get latest branch name, author of latest commit and time it was last updated.
If `package` has only `gh-pages` branches, a warning is issued and a single-row
data.frame filled with `NA` is returned invisibly.
}
\examples{
\dontrun{
get_latest_branch_update()
get_latest_branch_update("wbpip", display = FALSE)
}
}
