% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{get_latest_branch_update}
\alias{get_latest_branch_update}
\title{Get details of the branch which was last updated}
\usage{
get_latest_branch_update(package = "pipapi")
}
\arguments{
\item{package}{one of the core package (default "pipapi")}
}
\value{
single row dataframe
}
\description{
Get details of the branch which was last updated
}
\examples{
\dontrun{
get_latest_branch_update()
get_latest_branch_update("pipr")
}
}
