% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_metapip.R
\name{init_metapip}
\alias{init_metapip}
\title{Initializes the pip core packages}
\usage{
init_metapip()
}
\value{
invisible() output
}
\description{
Based on options() settings provides an option to download latest package versions from the branch
}
\examples{
\dontrun{
  init_metapip()
}

}
