% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_all_packages.R
\name{install_all_packages}
\alias{install_all_packages}
\title{Install all pip core packages from a specific branch.
This includes packages like pipapi, pipaux, pipload, wbpip, pipfun, pipdata and pipr}
\usage{
install_all_packages(branch = "PROD")
}
\arguments{
\item{branch}{valid branch name (default "PROD")}
}
\value{
invisible NULL
}
\description{
Install all pip core packages from a specific branch.
This includes packages like pipapi, pipaux, pipload, wbpip, pipfun, pipdata and pipr
}
\examples{
\dontrun{
install_all_packages("test")
}

}
