% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pip_packages.R
\name{install_branch}
\alias{install_branch}
\title{Install branch from a package}
\usage{
install_branch(package = "pipapi", branch = NULL)
}
\arguments{
\item{package}{one of the core package name (default "pipapi")}

\item{branch}{valid branch name (default "PROD")}
}
\description{
Install branch from a package
}
\examples{
\dontrun{
  install_branch()
  install_branch("pipfun", "ongoing")
}

}
