% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{install_branch}
\alias{install_branch}
\title{Install branch from a package}
\usage{
install_branch(package = "pipapi", branch = "master")
}
\arguments{
\item{package}{one of the core package name}

\item{branch}{valid branch name}
}
\description{
Install branch from a package
}
