% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_latest_branch.R
\name{install_latest_branch}
\alias{install_latest_branch}
\title{Install latest branch from core packages}
\usage{
install_latest_branch(package = NULL)
}
\arguments{
\item{package}{one (or more) of core packages. default NULL would install latest branch for all packages}
}
\description{
Install latest branch from core packages
}
\examples{
\dontrun{
  install_latest_branch()
  install_latest_branch(c("pipfun", "pipapi"))
}
}
