% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pip_packages.R
\name{install_pip_packages}
\alias{install_pip_packages}
\title{Install one (or more) pip core packages from a branch.}
\usage{
install_pip_packages(package = NULL, branch = "PROD")
}
\arguments{
\item{package}{one (or more) of the core package name, if NULL all the core packages are installed from the branch}

\item{branch}{valid branch name (default "PROD")}
}
\value{
invisible NULL
}
\description{
This installs packages like pipapi, pipaux, pipload, wbpip, pipfun, pipdata and pipr from a branch
}
\examples{
\dontrun{
install_pip_packages(branch = "test")
}

}
