% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapip-package.R
\docType{package}
\name{metapip-package}
\alias{metapip-package}
\title{metapip: Easily Install, Load, and manage PIP Packages}
\description{
This package helps you to install latest branch from PIP packages, get information about the packages and many more things. All of this is enabled via Github API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIP-Technical-Team/metapip}
  \item \url{https://pip-technical-team.github.io/metapip/}
  \item Report bugs at \url{https://github.com/PIP-Technical-Team/metapip/issues}
}

}
\author{
\strong{Maintainer}: R.Andrés Castañeda \email{acastanedaa@worldbank.org}

Authors:
\itemize{
  \item Ronak Shah \email{shahronak47@yahoo.in}
}

Other contributors:
\itemize{
  \item World Bank [copyright holder]
}

}
\keyword{internal}
