% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metapip_packages}
\alias{metapip_packages}
\title{List all metapip package dependencies}
\usage{
metapip_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. If `TRUE` (default), includes
`"metapip"` in the returned vector.}
}
\value{
Character vector of package names.
}
\description{
Returns the names of all packages listed in the `Imports` field of
metapip's DESCRIPTION. Optionally includes `"metapip"` itself.
}
\examples{
metapip_packages()
metapip_packages(include_self = FALSE)

}
