% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metapip_packages}
\alias{metapip_packages}
\title{List all packages in metapip}
\usage{
metapip_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metapip in the list?}
}
\description{
List all packages in metapip
}
\examples{
metapip_packages()
}
