% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{metapip_update}
\alias{metapip_update}
\title{Update metapip packages}
\usage{
metapip_update(pkg = "metapip", recursive = FALSE, ...)
}
\arguments{
\item{pkg}{A character string for the model being updated.}

\item{recursive}{If `TRUE`, will also check all dependencies of
metapip packages.}

\item{...}{Extra arguments to pass to [utils::install.packages()]}
}
\value{
Nothing is returned but a message is printed to the
 console about which packages (if any) should be installed along
 with code to do so.
}
\description{
This will check to see if all metapip packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
metapip_update()
}
}
