% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{package_branches}
\alias{package_branches}
\title{Status tables of package versions in different branches
Apart from PROD and DEV, we are only showing those branches whose version is higher than DEV.}
\usage{
package_branches(package = NULL)
}
\arguments{
\item{package}{One (or more) of the PIP core packages. Default NULL will include all the packages}
}
\value{
tibble of pipr packages and the corresponding package versions of branch
}
\description{
Status tables of package versions in different branches
Apart from PROD and DEV, we are only showing those branches whose version is higher than DEV.
}
\examples{
\dontrun{
package_branches()
package_branches(c("pipapi", "wbpip"))
}

}
