% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_branches.R
\name{package_branches}
\alias{package_branches}
\title{Status tables of package versions in different branches along with local installations.
For local installation,a status column is returned which indicates if the local version is ahead or behind the DEV branch.}
\usage{
package_branches(package = NULL)
}
\arguments{
\item{package}{One (or more) of the PIP core packages. Default NULL will include all the packages}
}
\value{
tibble of pip packages and the corresponding package versions of branch
}
\description{
Status tables of package versions in different branches along with local installations.
For local installation,a status column is returned which indicates if the local version is ahead or behind the DEV branch.
}
\examples{
\dontrun{
package_branches()
package_branches(c("pipapi", "wbpip"))
}

}
