% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pkg_deps}
\alias{pkg_deps}
\title{List all dependencies}
\usage{
pkg_deps(x = "metapip", recursive = FALSE)
}
\arguments{
\item{x}{A character string for the packages being evaluated.}

\item{recursive}{If `TRUE`, will also list all dependencies of
metapip packages.}
}
\description{
List all dependencies
}
