% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.metapip_simplelist}
\alias{print.metapip_simplelist}
\title{Pretty print for metapip_simplelist objects}
\usage{
\method{print}{metapip_simplelist}(x, ...)
}
\arguments{
\item{x}{A named list of character vectors (class 'metapip_simplelist').}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\description{
Prints a named list of character vectors (possibly empty) in a readable way using the cli package.
}
