% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowname_to_column}
\alias{rowname_to_column}
\title{Non tidyverse alternative to tibble::rownames_to_column}
\usage{
rowname_to_column(data, var)
}
\arguments{
\item{data}{Dataframe}

\item{var}{column name to store rownames}
}
\value{
Dataframe with an additional column of rownames
}
\description{
Non tidyverse alternative to tibble::rownames_to_column
}
\examples{
\dontrun{
rowname_to_column(mtcars, "rn")
}
}
