% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches.R
\name{set_custom_branch}
\alias{set_custom_branch}
\alias{get_custom_branch}
\title{Set default custom branching options}
\usage{
set_custom_branch(...)

get_custom_branch(package = NULL)
}
\arguments{
\item{...}{Named elements to be added or updated in the custom default list.}

\item{package}{character: vector with name of branches. E.g., c("pipdata",
"pipfaker"). Default return all packages whose default  branches have been
customed.}
}
\value{
invisible custom branches

Names list of packages and their corresponding branch
}
\description{
Set default custom branching options

Get Custom branches
}
\examples{
\dontrun{
set_custom_branch(pipr = 'main', 'pipapi' = 'DEV_v2')
}
get_custom_branch()
}
